Cleanup torrent lazily, but cleanup client immediately
- Fixes the "unknown message for clientKey" that happens in Brave when
navigating to another page. Used to happen because when the {delay:
5000} option was used, the client wasn't removed from the torrent right
away, which caused 'update' messages to continue getting sent.

- I removed the `delay` option, now the client is always removed right
away from the torrent. But the torrent isn't destroyed for 10s, giving
time for another client to connect. If a client connects in that time,
the torrent is not destroyed.

- This is nice, because now it matches the WebTorrent API more closely,
which doesn't have a `delay` option.

- Refactored `killClients()` to just kill a single client `killClient()`